2927: Refactored poster subscription
